feat(storage): auto-create artifact when not found during run
When `--artifact-name` references a non-existent artifact, automatically
create the storage record and an empty initial version instead of failing.
This eliminates the need for manual `vm0 artifact init` + `vm0 artifact push`
before the first run.

Extracts the artifact creation logic from Slack's `ensureScopeAndArtifact()`
into a shared `ensureArtifactExists()` function, removing ~100 lines of
duplicated code.

Closes #3413
